Autonomous and Sensors added
We now have auto and sensors and xbox controller support

+++ src/CCTalon/CCTalon.java
@@ -11,7 +11,7 @@ public class CCTalon extends Talon
 		_reverse = reverse;
 	}
 	
-	public void set(int speed) {
+	public void set(double speed) {
 		if (_reverse) 
 		{
 			super.set(-speed);
